requirements.txt
anthropic
python-dotenv
requests
beautifulsoup4
weasyprint
jinja2
